Change type of product.prices from NSArray to NSMutableArray in order to add new price on product retain object when a new price is posted
git-svn-id: https://scan-and-share.googlecode.com/svn/trunk@46 10737fd6-5b37-671e-7c70-df2a59443892

+++ Scan&Share/Scan&Share/Classes/SSProduct.m
@@ -19,7 +19,7 @@
         SSPrice *min = [self.prices objectAtIndex:0];
         
         for (SSPrice *p in self.prices) {
-            if (p.value < min.value){
+            if (p.value.floatValue < min.value.floatValue){
                 min = p;
             }
         }
@@ -35,7 +35,7 @@
         SSPrice *max = [self.prices objectAtIndex:0];
     
         for (SSPrice *p in self.prices) {
-            if (p.value > max.value){
+            if (p.value.floatValue > max.value.floatValue){
                 max = p;
             }
         }
